requirements.txt
streamlit
pymongo
google-generativeai
python-dotenv
dnspython